Reverted NativeOrAssetId
